Bug 1010163 - Cast out warning demons on a CLOSED TREE.
UltraBlame original commit: 72230d707f866b451d4a784514fe70c0f655a3c0

+++ content/media/wave/WaveReader.cpp
@@ -577,7 +577,7 @@ WaveReader::LoadListChunk(uint32_t aChunkSize,
     uint32_t length = ReadUint32LE(&p);
 
     
-    if (end - p < length) {
+    if (uint32_t(end - p) < length) {
       break;
     }
